refactor: utils -> provider 패키지명 변경
